pyproject.toml
[project]
name = "thresher"
version = "0.1.0"
description = "Cloud native document extraction and indexing pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chonkie[code]>=1.3.0",
    "docling[asr]>=2.0.0",
    "docling-core[chunking]",
    "google-cloud-storage>=2.0.0",
    "jsonschema>=4.26.0",
    "kubernetes>=28.0.0",
    "python-magic>=0.4.27",
    "pyyaml>=6.0",
    "qdrant-client[fastembed]>=1.12.0",
]

[project.scripts]
thresher = "thresher.cli:main"

[tool.setuptools.packages.find]
include = ["thresher*"]

[tool.setuptools.package-data]
thresher = ["defaults.yaml", "config_schema.json"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (GCS, Qdrant)",
    "contract: marks provider contract tests",
    "functional: marks functional tests requiring Docker services (fake-gcs-server, qdrant)",
]

[dependency-groups]
dev = [
    "prek>=0.3.8",
    "pytest>=8.0",
    "pytest-timeout>=2.3.1",
    "requests>=2.32.5",
    "ruff>=0.15.7",
    "ty>=0.0.25",
]
